support nested queries
