+++ package.json
@@ -30,7 +30,8 @@
     "Nauval Atmaja",
     "Ferron Hanse",
     "Dave Cranwell",
-    "Anton Samper Rivaya"
+    "Anton Samper Rivaya",
+    "Johannes Scharlach"
   ],
   "license": "MIT",
   "devDependencies": {
